pyproject.toml
[project]
name = "nws-monitor"
version = "0.1.0"
description = "NOAA Weather Radio monitor using ka9q-radio"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "websockets",
    "maidenhead",
    "httpx",
    "numpy>=1.24.0",
    "ka9q-python>=3.4.2",
]

[tool.setuptools]
packages = ["backend"]
